+++ packages/sitecore-jss/src/editing/models.ts
@@ -1,0 +1,9 @@
+﻿/**
+ * Represents the Editing Layout variant.
+ * - shared - shared layout
+ * - final - final layout
+ */
+export enum LayoutKind {
+  Final = 'final',
+  Shared = 'shared',
+}