+++ CHANGELOG.md
@@ -55,7 +55,10 @@ Our versioning strategy is as follows:
 
 ### 🎉 New Features & Improvements
 
+NOTE: API was changed, preview data provides a new parameter therefore, please make the necessary updates to the nextjs template files as shown in the following PR to experience a smooth upgrade.
+
 * `[sitecore-jss]``[sitecore-jss-nextjs]` Pass `sc_layoutKind` to GraphQLEditingService request header to support shared/final editing layouts ([#1907](https://github.com/Sitecore/jss/pull/1907))
+  * NOTE: `GraphQLRequestClient`'s  `request` method now supports dynamic headers.
 
 ## 22.1.0